Move value classes and update their imports in all the places
