+++ src/Tickets/Commerce/Cart/Abstract_Cart.php
@@ -11,8 +11,8 @@ namespace TEC\Tickets\Commerce\Cart;
 
 use InvalidArgumentException;
 use TEC\Tickets\Commerce\Cart;
-use TEC\Tickets\Commerce\Order_Modifiers\Values\Legacy_Value_Factory as Factory;
-use TEC\Tickets\Commerce\Order_Modifiers\Values\Precision_Value;
+use TEC\Tickets\Commerce\Values\Legacy_Value_Factory as Factory;
+use TEC\Tickets\Commerce\Values\Precision_Value;
 use TEC\Tickets\Commerce\Traits\Cart as Cart_Trait;
 use TEC\Tickets\Commerce\Utils\Value;
 use Tribe__Tickets__Tickets as Tickets;
