+++ src/Tickets/Commerce/Order_Modifiers/API/Coupons.php
@@ -18,7 +18,7 @@ use TEC\Tickets\Commerce\Order_Modifiers\Models\Order_Modifier;
 use TEC\Tickets\Commerce\Order_Modifiers\Modifiers\Coupon_Modifier_Manager as Manager;
 use TEC\Tickets\Commerce\Order_Modifiers\Repositories\Coupons as Coupons_Repository;
 use TEC\Tickets\Commerce\Order_Modifiers\Traits\Coupons as CouponsTrait;
-use TEC\Tickets\Commerce\Order_Modifiers\Values\Currency_Value;
+use TEC\Tickets\Commerce\Values\Currency_Value;
 use TEC\Tickets\Commerce\Traits\Type;
 use WP_Error;
 use WP_REST_Request as Request;
