+++ src/Tickets/Commerce/Order_Modifiers/Modifiers/Coupon.php
@@ -14,7 +14,7 @@ namespace TEC\Tickets\Commerce\Order_Modifiers\Modifiers;
 
 use TEC\Common\StellarWP\Models\Contracts\Model;
 use TEC\Tickets\Commerce\Order_Modifiers\Table_Views\Coupon_Table;
-use TEC\Tickets\Commerce\Order_Modifiers\Values\Precision_Value;
+use TEC\Tickets\Commerce\Values\Precision_Value;
 use Tribe__Tickets__Admin__Views;
 
 /**
